add  modal for mobile filter + refactoring
