+++ src/components/catalog-page/catalog-page.tsx
@@ -1,7 +1,7 @@
 import { getCatalog } from '@/api/api';
 import Card from '@/components/card/Card';
 import CatalogFilter from './catalog-filter';
-import FilterMobile from './filter-mobile';
+
 
 const CatalogPage = async () => {
   const cardsCatalog = await getCatalog();
@@ -13,7 +13,7 @@ const CatalogPage = async () => {
   return (
     <>
       <CatalogFilter cardsCatalog={cardsCatalog}/>
-      {/* <FilterMobile/> */}
+      
       {/* <section className="flex items-center flex-col  my-5 md:my-9 md:px-20 md:flex-row md:flex-wrap justify-between lg:justify-evenly xl:justify-between md:gap-[76px] lg:gap-5 lg:px-[164px]">
         {cardsCatalog.map(card => (
           <Card key={card.id} card={card} />
